Rename, renamespace Request\PayloadStore
