+++ src/Controller/FileSourceFileController.php
@@ -7,9 +7,9 @@ namespace App\Controller;
 use App\Enum\ApiService;
 use App\Exception\ApiException;
 use App\Request\FileSourceFileRequest;
-use App\Request\PayloadStore;
 use App\Response\RedirectResponse;
 use App\Security\ApiKey;
+use App\SessionStore\RequestPayloadStore;
 use SmartAssert\ApiClient\Exception\ClientException;
 use SmartAssert\ApiClient\FileClient;
 use Symfony\Component\HttpFoundation\Response;
@@ -32,8 +32,11 @@ readonly class FileSourceFileController
         name: 'sources_create_file_source_file',
         methods: ['POST']
     )]
-    public function create(ApiKey $apiKey, FileSourceFileRequest $request, PayloadStore $payloadStore): Response
-    {
+    public function create(
+        ApiKey $apiKey,
+        FileSourceFileRequest $request,
+        RequestPayloadStore $requestPayloadStore
+    ): Response {
         $response = new RedirectResponse(
             $this->urlGenerator->generate('sources_view_file_source', ['id' => $request->sourceId])
         );
@@ -41,7 +44,7 @@ readonly class FileSourceFileController
         try {
             $this->fileClient->create($apiKey->key, $request->sourceId, $request->filename, $request->content);
         } catch (ClientException $e) {
-            $payloadStore->set($request);
+            $requestPayloadStore->set($request);
 
             throw new ApiException(ApiService::SOURCES, $e, $response);
         }
